+++ src/lightning/pytorch/trainer/trainer.py
@@ -37,7 +37,6 @@ import torch.distributed as dist
 from lightning_utilities.core.apply_func import apply_to_collection
 from lightning_utilities.core.imports import module_available
 from packaging.version import Version
-from torch import Tensor
 from torch.optim import Optimizer
 from torch.utils.data import DataLoader
 
@@ -45,7 +44,6 @@ import lightning.pytorch as pl
 from lightning.fabric.utilities.apply_func import convert_tensors_to_scalars
 from lightning.fabric.utilities.cloud_io import get_filesystem
 from lightning.fabric.utilities.data import _auto_add_worker_init_fn
-from lightning.fabric.utilities.imports import _TORCH_GREATER_EQUAL_2_0
 from lightning.fabric.utilities.types import _PATH
 from lightning.fabric.utilities.warnings import PossibleUserWarning
 from lightning.pytorch.accelerators import Accelerator, TPUAccelerator
@@ -54,13 +52,14 @@ from lightning.pytorch.callbacks.prediction_writer import BasePredictionWriter
 from lightning.pytorch.core.datamodule import LightningDataModule
 from lightning.pytorch.loggers import Logger
 from lightning.pytorch.loggers.tensorboard import TensorBoardLogger
+from lightning.pytorch.loggers.utilities import _log_hyperparams
 from lightning.pytorch.loops import _PredictionLoop, _TrainingEpochLoop
 from lightning.pytorch.loops.dataloader.evaluation_loop import _EvaluationLoop
 from lightning.pytorch.loops.fit_loop import _FitLoop
 from lightning.pytorch.loops.utilities import _parse_loop_limits, _reset_progress
 from lightning.pytorch.plugins import PLUGIN_INPUT, PrecisionPlugin
 from lightning.pytorch.profilers import Profiler
-from lightning.pytorch.strategies import DDPStrategy, FSDPStrategy, ParallelStrategy, SingleDeviceStrategy, Strategy
+from lightning.pytorch.strategies import ParallelStrategy, Strategy
 from lightning.pytorch.trainer import call, setup
 from lightning.pytorch.trainer.configuration_validator import verify_loop_configurations
 from lightning.pytorch.trainer.connectors.accelerator_connector import (
@@ -85,6 +84,7 @@ from lightning.pytorch.utilities.argparse import (
     parse_argparser,
     parse_env_variables,
 )
+from lightning.pytorch.utilities.compile import _maybe_unwrap_optimized, _verify_strategy_supports_compile
 from lightning.pytorch.utilities.data import has_len_all_ranks
 from lightning.pytorch.utilities.exceptions import MisconfigurationException
 from lightning.pytorch.utilities.model_helpers import is_overridden
@@ -308,8 +308,7 @@ class Trainer:
                 evaluation (``validate``/``test``/``predict``).
         """
         super().__init__()
-        Trainer._log_api_event("init")
-        log.detail(f"{self.__class__.__name__}: Initializing trainer with parameters: {locals()}")
+        log.debug(f"{self.__class__.__name__}: Initializing trainer with parameters: {locals()}")
         self.state = TrainerState()
 
         if default_root_dir is not None:
@@ -444,21 +443,6 @@ class Trainer:
         self._last_train_dl_reload_epoch = float("-inf")
         self._last_val_dl_reload_epoch = float("-inf")
 
-    def _maybe_unwrap_optimized(self, model: object) -> "pl.LightningModule":
-        if not _TORCH_GREATER_EQUAL_2_0:
-            if not isinstance(model, pl.LightningModule):
-                raise TypeError(f"`model` must be a `LightningModule`, got `{type(model).__qualname__}`")
-            return model
-        from torch._dynamo import OptimizedModule
-
-        if isinstance(model, OptimizedModule):
-            return model.from_compiled(model)
-        if isinstance(model, pl.LightningModule):
-            return model
-        raise TypeError(
-            f"`model` must be a `LightningModule` or `torch._dynamo.OptimizedModule`, got `{type(model).__qualname__}`"
-        )
-
     def fit(
         self,
         model: "pl.LightningModule",
@@ -485,7 +469,7 @@ class Trainer:
 
             datamodule: An instance of :class:`~lightning.pytorch.core.datamodule.LightningDataModule`.
         """
-        model = self._maybe_unwrap_optimized(model)
+        model = _maybe_unwrap_optimized(model)
         self.strategy._lightning_module = model
         call._call_and_handle_interrupt(
             self, self._fit_impl, model, train_dataloaders, val_dataloaders, datamodule, ckpt_path
@@ -499,8 +483,7 @@ class Trainer:
         datamodule: Optional[LightningDataModule] = None,
         ckpt_path: Optional[str] = None,
     ) -> None:
-        Trainer._log_api_event("fit")
-        log.detail(f"{self.__class__.__name__}: trainer fit stage")
+        log.debug(f"{self.__class__.__name__}: trainer fit stage")
 
         self.state.fn = TrainerFn.FITTING
         self.state.status = TrainerStatus.RUNNING
@@ -571,7 +554,7 @@ class Trainer:
                     "`Trainer.validate()` requires a `LightningModule` when it hasn't been passed in a previous run"
                 )
         else:
-            model = self._maybe_unwrap_optimized(model)
+            model = _maybe_unwrap_optimized(model)
             self.strategy._lightning_module = model
         return call._call_and_handle_interrupt(
             self, self._validate_impl, model, dataloaders, ckpt_path, verbose, datamodule
@@ -588,8 +571,7 @@ class Trainer:
         # --------------------
         # SETUP HOOK
         # --------------------
-        Trainer._log_api_event("validate")
-        log.detail(f"{self.__class__.__name__}: trainer validate stage")
+        log.debug(f"{self.__class__.__name__}: trainer validate stage")
 
         self.state.fn = TrainerFn.VALIDATING
         self.state.status = TrainerStatus.RUNNING
@@ -663,7 +645,7 @@ class Trainer:
                     "`Trainer.test()` requires a `LightningModule` when it hasn't been passed in a previous run"
                 )
         else:
-            model = self._maybe_unwrap_optimized(model)
+            model = _maybe_unwrap_optimized(model)
             self.strategy._lightning_module = model
         return call._call_and_handle_interrupt(
             self, self._test_impl, model, dataloaders, ckpt_path, verbose, datamodule
@@ -680,8 +662,7 @@ class Trainer:
         # --------------------
         # SETUP HOOK
         # --------------------
-        Trainer._log_api_event("test")
-        log.detail(f"{self.__class__.__name__}: trainer test stage")
+        log.debug(f"{self.__class__.__name__}: trainer test stage")
 
         self.state.fn = TrainerFn.TESTING
         self.state.status = TrainerStatus.RUNNING
@@ -757,7 +738,7 @@ class Trainer:
                     "`Trainer.predict()` requires a `LightningModule` when it hasn't been passed in a previous run"
                 )
         else:
-            model = self._maybe_unwrap_optimized(model)
+            model = _maybe_unwrap_optimized(model)
             self.strategy._lightning_module = model
         return call._call_and_handle_interrupt(
             self, self._predict_impl, model, dataloaders, datamodule, return_predictions, ckpt_path
@@ -774,8 +755,7 @@ class Trainer:
         # --------------------
         # SETUP HOOK
         # --------------------
-        Trainer._log_api_event("predict")
-        log.detail(f"{self.__class__.__name__}: trainer predict stage")
+        log.debug(f"{self.__class__.__name__}: trainer predict stage")
 
         self.state.fn = TrainerFn.PREDICTING
         self.state.status = TrainerStatus.RUNNING
@@ -812,17 +792,7 @@ class Trainer:
     def _run(
         self, model: "pl.LightningModule", ckpt_path: Optional[_PATH] = None
     ) -> Optional[Union[_EVALUATE_OUTPUT, _PREDICT_OUTPUT]]:
-        if model._compiler_ctx is not None:
-            supported_strategies = [SingleDeviceStrategy, DDPStrategy, FSDPStrategy]
-            if self.strategy is not None and not any(isinstance(self.strategy, s) for s in supported_strategies):
-                supported_strategy_names = ", ".join(s.__name__ for s in supported_strategies)
-                raise RuntimeError(
-                    "Using a compiled model is incompatible with the current strategy: "
-                    f"{self.strategy.__class__.__name__}. "
-                    f"Only {supported_strategy_names} support compilation. "
-                    "Either switch to one of the supported strategies or avoid passing in "
-                    "a compiled model."
-                )
+        _verify_strategy_supports_compile(model, self.strategy)
 
         if self.state.fn == TrainerFn.FITTING:
             min_epochs, max_epochs = _parse_loop_limits(
@@ -844,13 +814,13 @@ class Trainer:
         verify_loop_configurations(self)
 
         # hook
-        log.detail(f"{self.__class__.__name__}: preparing data")
+        log.debug(f"{self.__class__.__name__}: preparing data")
         self._data_connector.prepare_data()
 
         # ----------------------------
         # SET UP TRAINING
         # ----------------------------
-        log.detail(f"{self.__class__.__name__}: setting up strategy environment")
+        log.debug(f"{self.__class__.__name__}: setting up strategy environment")
         self.strategy.setup_environment()
         self.__setup_profiler()
 
@@ -858,10 +828,10 @@ class Trainer:
 
         # check if we should delay restoring checkpoint till later
         if not self.strategy.restore_checkpoint_after_setup:
-            log.detail(f"{self.__class__.__name__}: restoring module and callbacks from checkpoint path: {ckpt_path}")
+            log.debug(f"{self.__class__.__name__}: restoring module and callbacks from checkpoint path: {ckpt_path}")
             self._checkpoint_connector._restore_modules_and_callbacks(ckpt_path)
 
-        log.detail(f"{self.__class__.__name__}: configuring sharded model")
+        log.debug(f"{self.__class__.__name__}: configuring sharded model")
         self._call_configure_sharded_model()  # allow user to setup in model sharded environment
 
         # ----------------------------
@@ -903,21 +873,21 @@ class Trainer:
             self._call_callback_hooks("on_fit_start")
             self._call_lightning_module_hook("on_fit_start")
 
-        self._log_hyperparams()
+        _log_hyperparams(self)
 
         if self.strategy.restore_checkpoint_after_setup:
-            log.detail(f"{self.__class__.__name__}: restoring module and callbacks from checkpoint path: {ckpt_path}")
+            log.debug(f"{self.__class__.__name__}: restoring module and callbacks from checkpoint path: {ckpt_path}")
             self._checkpoint_connector._restore_modules_and_callbacks(ckpt_path)
 
         # restore optimizers, etc.
-        log.detail(f"{self.__class__.__name__}: restoring training state")
+        log.debug(f"{self.__class__.__name__}: restoring training state")
         self._checkpoint_connector.restore_training_state()
 
         self._checkpoint_connector.resume_end()
 
         results = self._run_stage()
 
-        log.detail(f"{self.__class__.__name__}: trainer tearing down")
+        log.debug(f"{self.__class__.__name__}: trainer tearing down")
         self._teardown()
 
         # ----------------------------
@@ -928,52 +898,13 @@ class Trainer:
             self._call_callback_hooks("on_fit_end")
             self._call_lightning_module_hook("on_fit_end")
 
-        log.detail(f"{self.__class__.__name__}: calling teardown hooks")
+        log.debug(f"{self.__class__.__name__}: calling teardown hooks")
         self._call_teardown_hook()
 
         self.state.status = TrainerStatus.FINISHED
         self.state.stage = None
 
         return results
-
-    def _log_hyperparams(self) -> None:
-        if not self.loggers:
-            return
-        # log hyper-parameters
-        hparams_initial = None
-
-        # save exp to get started (this is where the first experiment logs are written)
-        datamodule_log_hyperparams = self.datamodule._log_hyperparams if self.datamodule is not None else False
-
-        if self.lightning_module._log_hyperparams and datamodule_log_hyperparams:
-            datamodule_hparams = self.datamodule.hparams_initial
-            lightning_hparams = self.lightning_module.hparams_initial
-            inconsistent_keys = []
-            for key in lightning_hparams.keys() & datamodule_hparams.keys():
-                lm_val, dm_val = lightning_hparams[key], datamodule_hparams[key]
-                if type(lm_val) != type(dm_val):
-                    inconsistent_keys.append(key)
-                elif isinstance(lm_val, Tensor) and id(lm_val) != id(dm_val):
-                    inconsistent_keys.append(key)
-                elif lm_val != dm_val:
-                    inconsistent_keys.append(key)
-            if inconsistent_keys:
-                raise MisconfigurationException(
-                    f"Error while merging hparams: the keys {inconsistent_keys} are present "
-                    "in both the LightningModule's and LightningDataModule's hparams "
-                    "but have different values."
-                )
-            hparams_initial = {**lightning_hparams, **datamodule_hparams}
-        elif self.lightning_module._log_hyperparams:
-            hparams_initial = self.lightning_module.hparams_initial
-        elif datamodule_log_hyperparams:
-            hparams_initial = self.datamodule.hparams_initial
-
-        for logger in self.loggers:
-            if hparams_initial is not None:
-                logger.log_hyperparams(hparams_initial)
-            logger.log_graph(self.lightning_module)
-            logger.save()
 
     def _teardown(self) -> None:
         """This is the Trainer's internal teardown, unrelated to the `teardown` hooks in LightningModule and
@@ -1117,8 +1048,6 @@ class Trainer:
         # these could have become stale if metrics are defined in `setup`
         self.lightning_module._metric_attributes = None
 
-        # todo: TPU 8 cores hangs in flush with TensorBoard. Might do for all loggers.
-        # It might be related to xla tensors blocked when moving the cpu kill loggers.
         for logger in self.loggers:
             logger.finalize("success")
 
@@ -1216,14 +1145,7 @@ class Trainer:
 
         for callback in self.callbacks:
             with self.profiler.profile(f"[Callback]{callback.state_key}.on_save_checkpoint"):
-                state = callback.on_save_checkpoint(self, self.lightning_module, checkpoint)
-            if state is not None:
-                # TODO: Remove this error message in v2.0
-                raise ValueError(
-                    f"Returning a value from `{callback.__class__.__name__}.on_save_checkpoint` was deprecated in v1.6"
-                    f" and is no longer supported as of v1.8. Please override `Callback.state_dict` to return state"
-                    f" to be saved."
-                )
+                callback.on_save_checkpoint(self, self.lightning_module, checkpoint)
 
         if pl_module:
             # restore current_fx when nested context
@@ -1297,10 +1219,6 @@ class Trainer:
         pl_module._current_fx_name = prev_fx_name
 
         return output
-
-    @staticmethod
-    def _log_api_event(event: str) -> None:
-        torch._C._log_api_usage_once("lightning.trainer." + event)
 
     def __setup_profiler(self) -> None:
         assert self.state.fn is not None
@@ -1599,10 +1517,6 @@ class Trainer:
             return self.strategy.distributed_sampler_kwargs
 
     @property
-    def data_parallel(self) -> bool:
-        return isinstance(self.strategy, ParallelStrategy)
-
-    @property
     def enable_validation(self) -> bool:
         """Check if we should run validation during training."""
         return (
@@ -1888,9 +1802,7 @@ class Trainer:
         self._loggers = loggers if loggers else []
 
     @property
-    def callback_metrics(self) -> Dict:
-        # TODO: the true typing return can include dictionaries as defined in
-        # `lightning.pytorch.trainer.connectors.logger_connector.result._OUT_DICT`
+    def callback_metrics(self) -> _OUT_DICT:
         return self._logger_connector.callback_metrics
 
     @property